requirements.txt
django
django-grappelli
django-allauth
python-dotenv
qrcode
Pillow
gunicorn
psycopg2-binary
dj-database-url

